chore: rename mapSearch to mapSearchTools
